added fix for storage fee computation

+++ database/migrations/2018_09_02_085152_add_balance_on_users_table.php
@@ -14,7 +14,7 @@ class AddBalanceOnUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('balance')->default(0);
+            $table->double('balance', 8, 2)->default(0);
         });
     }
